fix: ksp with arch filter
