+++ build-logic/src/main/kotlin/io/github/droidkaigi/confsched2023/primitive/KmpIosPlugin.kt
@@ -8,21 +8,12 @@ import org.gradle.api.Project
 import org.gradle.kotlin.dsl.get
 import org.gradle.kotlin.dsl.withType
 import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget
-import java.util.Properties
 
 @Suppress("unused")
 class KmpIosPlugin : Plugin<Project> {
     override fun apply(target: Project) {
         with(target) {
             kotlin {
-                val activeArch = Arch.findByArch(
-                    rootProject.layout.projectDirectory.file("local.properties").asFile.takeIf { it.exists() }
-                        ?.let {
-                            Properties().apply {
-                                load(it.reader(Charsets.UTF_8))
-                            }.getProperty("arch")
-                        } ?: System.getenv("arch")
-                )
                 when (activeArch) {
                     ARM -> iosSimulatorArm64()
                     X86 -> iosX64()
@@ -33,25 +24,18 @@ class KmpIosPlugin : Plugin<Project> {
                     }
                 }
                 with(sourceSets) {
-                    val iosSourceSets = listOf(
-                        "iosArm64",
-                        "iosX64",
-                        "iosSimulatorArm64",
-                    )
                     create("iosMain") {
                         dependsOn(getByName("commonMain"))
-                    }
-                    iosSourceSets.forEach { iosSourceSet ->
-                        maybeCreate(iosSourceSet + "Main")
-                            .dependsOn(getByName("iosMain"))
+                        maybeCreate("iosArm64Main").dependsOn(this)
+                        maybeCreate("iosX64Main").dependsOn(this)
+                        maybeCreate("iosSimulatorArm64Main").dependsOn(this)
                     }
 
                     create("iosTest") {
                         dependsOn(getByName("commonTest"))
-                    }
-                    iosSourceSets.forEach { iosSourceSet ->
-                        maybeCreate(iosSourceSet + "Test")
-                            .dependsOn(getByName("iosTest"))
+                        maybeCreate("iosArm64Test").dependsOn(this)
+                        maybeCreate("iosX64Test").dependsOn(this)
+                        maybeCreate("iosSimulatorArm64Test").dependsOn(this)
                     }
                 }
 
@@ -61,19 +45,6 @@ class KmpIosPlugin : Plugin<Project> {
                     compilations["main"].kotlinOptions.freeCompilerArgs += "-Xexport-kdoc"
                 }
             }
-        }
-    }
-}
-
-enum class Arch(val arch: String?) {
-    ARM("arm64"),
-    X86("x86_64"),
-    ALL(null),
-    ;
-
-    companion object {
-        fun findByArch(arch: String?): Arch {
-            return values().firstOrNull { it.arch == arch } ?: ALL
         }
     }
 }
