+++ build-logic/src/main/kotlin/io/github/droidkaigi/confsched2023/primitive/KmpKtorfitPlugin.kt
@@ -1,11 +1,13 @@
 package io.github.droidkaigi.confsched2023.primitive
 
+import io.github.droidkaigi.confsched2023.primitive.Arch.ALL
+import io.github.droidkaigi.confsched2023.primitive.Arch.ARM
+import io.github.droidkaigi.confsched2023.primitive.Arch.X86
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.dependencies
 import org.gradle.kotlin.dsl.get
-import org.gradle.kotlin.dsl.getByName
 
 @Suppress("unused")
 class KmpKtorfitPlugin : Plugin<Project> {
@@ -16,11 +18,11 @@ class KmpKtorfitPlugin : Plugin<Project> {
                 apply("de.jensklingenberg.ktorfit")
             }
             kotlin {
-                android()
+                androidTarget()
             }
 
             configure<de.jensklingenberg.ktorfit.gradle.KtorfitGradleConfiguration> {
-                version = "1.4.1"
+                version = "1.5.0"
             }
 
             kotlin {
@@ -33,16 +35,20 @@ class KmpKtorfitPlugin : Plugin<Project> {
             dependencies {
                 this.add(
                     "kspCommonMainMetadata",
-                    "de.jensklingenberg.ktorfit:ktorfit-ksp:1.4.1"
+                    "de.jensklingenberg.ktorfit:ktorfit-ksp:1.5.0"
                 )
-                this.add("kspAndroid", "de.jensklingenberg.ktorfit:ktorfit-ksp:1.4.1")
-                val iosSourceSets = listOf(
-                    "IosArm64",
-                    "IosX64",
-                    "IosSimulatorArm64",
-                )
-                iosSourceSets.forEach {
-                    this.add("ksp$it", "de.jensklingenberg.ktorfit:ktorfit-ksp:1.4.1")
+                this.add("kspAndroid", "de.jensklingenberg.ktorfit:ktorfit-ksp:1.5.0")
+                val iosConfigs = when (activeArch) {
+                    ARM -> listOf("IosSimulatorArm64")
+                    X86 -> listOf("IosX64")
+                    ALL -> listOf(
+                        "IosArm64",
+                        "IosX64",
+                        "IosSimulatorArm64",
+                    )
+                }
+                iosConfigs.forEach {
+                    this.add("ksp$it", "de.jensklingenberg.ktorfit:ktorfit-ksp:1.5.0")
                 }
             }
         }
